+++ limit-core/src/main/java/com/athena/module/sequence/mapper/SequenceMapper.java
@@ -1,9 +1,11 @@
 package com.athena.module.sequence.mapper;
 
+import java.math.BigDecimal;
+
 import org.apache.ibatis.annotations.Param;
 
 public interface SequenceMapper {
 
-	Long next(@Param("sequence") String sequence);
+	BigDecimal next(@Param("sequence") String sequence);
 
 }
